A New Beginning — Update script to 2.11 syntax

+++ play.it-2/games/play-a-new-beginning.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -31,47 +31,49 @@ set -o errexit
 
 ###
 # A New Beginning
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180802.1
+script_version=20190311.1
 
 # Set game-specific variables
 
 GAME_ID='a-new-beginning'
 GAME_NAME='A New Beginning'
 
-ARCHIVE_GOG='setup_a_new_beginning_final_cut_2.2.0.7.bin'
+ARCHIVE_GOG='setup_a_new_beginning_final_cut_2.2.0.7.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/a_new_beginning'
-ARCHIVE_GOG_MD5='1babc74ae7f29ff6ce16ea9fd3e4d3ff'
+ARCHIVE_GOG_MD5='7f0acf99bfbaabb46dfcba12340098df'
 ARCHIVE_GOG_SIZE='4100000'
 ARCHIVE_GOG_VERSION='1.0-gog2.2.0.7'
 ARCHIVE_GOG_TYPE='rar'
 ARCHIVE_GOG_GOGID='1207659150'
+ARCHIVE_GOG_PART1='setup_a_new_beginning_final_cut_2.2.0.7.bin'
+ARCHIVE_GOG_PART1_MD5='1babc74ae7f29ff6ce16ea9fd3e4d3ff'
+ARCHIVE_GOG_PART1_TYPE='rar'
 
 ARCHIVE_DOC_DATA_PATH='game'
-ARCHIVE_DOC_DATA_FILES='./*.pdf'
+ARCHIVE_DOC_DATA_FILES='*.pdf'
 
 ARCHIVE_GAME0_BIN_PATH='game'
-ARCHIVE_GAME0_BIN_FILES='./anb.exe ./VisionaireConfigurationTool.exe ./avcodec-54.dll ./avformat-54.dll ./avutil-52.dll ./libsndfile-1.dll ./openal32.dll ./SDL2.dll ./swresample-0.dll ./swscale-2.dll ./zlib1.dll'
+ARCHIVE_GAME0_BIN_FILES='anb.exe VisionaireConfigurationTool.exe avcodec-54.dll avformat-54.dll avutil-52.dll libsndfile-1.dll openal32.dll SDL2.dll swresample-0.dll swscale-2.dll zlib1.dll'
 
 ARCHIVE_GAME1_BIN_PATH='support/app'
-ARCHIVE_GAME1_BIN_FILES='./config.ini'
+ARCHIVE_GAME1_BIN_FILES='config.ini'
 
 ARCHIVE_GAME_DATA_PATH='game'
-ARCHIVE_GAME_DATA_FILES='./banner.jpg ./characters ./data.vis ./documents ./folder.jpg ./Languages.xml ./scenes ./videos'
+ARCHIVE_GAME_DATA_FILES='banner.jpg characters data.vis documents folder.jpg Languages.xml scenes videos'
 
 CONFIG_FILES='./config.ini'
 
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='anb.exe'
 APP_MAIN_ICON='anb.exe'
-APP_MAIN_ICON_RES='16 32 48'
 
 APP_WINETRICKS='dotnet40'
 
-PACKAGES_LIST=' PKG_DATA PKG_BIN'
+PACKAGES_LIST='PKG_DATA PKG_BIN'
 
 PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
@@ -81,39 +83,48 @@ PKG_BIN_DEPS="$PKG_DATA_ID wine winetricks"
 
 # Load common functions
 
-target_version='2.7'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
 prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Extract icons
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Store saved games outside of WINE prefix
 
@@ -127,7 +138,6 @@ pattern="$pattern\\tln --symbolic \"\$PATH_DATA/saves\" \"$save_path\"#"
 for file in "${PKG_BIN_PATH}${PATH_BIN}"/*; do
 	sed --in-place "$pattern" "$file"
 done
-
 
 # Build package
